+++ crates/swc_ecma_transforms_base/src/fixer.rs
@@ -100,9 +100,13 @@ impl Fixer<'_> {
     fn wrap_callee(&mut self, e: &mut Expr) {
         if match e {
             Expr::Lit(Lit::Num(..) | Lit::Str(..)) => false,
-            Expr::Cond(..) | Expr::Bin(..) | Expr::Lit(..) | Expr::Unary(..) | Expr::Object(..) => {
-                true
-            }
+            Expr::Cond(..)
+            | Expr::Bin(..)
+            | Expr::Lit(..)
+            | Expr::Unary(..)
+            | Expr::Object(..)
+            | Expr::Await(..)
+            | Expr::Yield(..) => true,
             _ => false,
         } {
             self.wrap(e)
@@ -1622,4 +1626,14 @@ var store = global[SHARED] || (global[SHARED] = {});
     identical!(issue_5109_1, "(0, b)?.()");
     identical!(issue_5109_2, "1 + (0, b)?.()");
     identical!(issue_5109_3, "(0, a)() ? undefined : (0, b)?.()");
+
+    identical!(
+        issue_5313,
+        "
+        async function* foo() {
+            (await a)();
+            (yield b)();
+        }
+        "
+    );
 }